backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
gunicorn==21.2.0
whitenoise==6.6.0
